endterm/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
click==8.1.7
dnspython==2.6.1
dramatiq==1.16.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
kafka-python==2.0.2
Mako==1.3.2
MarkupSafe==2.1.5
prometheus_client==0.20.0
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
redis==5.0.4
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
